D3CTF2019/bet2loss_v2/bet2loss_source/requirements.txt
Django
django-crontab==0.7.1
pysha3==1.0.2
web3==5.2.2
gunicorn==19.9.0
whitenoise==3.3.1
